Beginning of enemy movement
Bad guy added in (not final sprite, i found a new one)
I got drunk, came home and coded. It doesn't work as it should
But you know, it's a framework for what I'll fix tomorrow.

Note to self: fix the time delay/step for how often he moves
fix it so that once it hits the edge it reverses direction

Note to partners in the project: whiskey is delicious.