Added in Enemies
I added in enemies into the BattleMaster script. They should just work
with them now. I changed the Reset method to take 3 int now for # of
physical, magical, and boss mobs now instead of just # of enemies. Also
the enemies are currently roughly balanced so that 1 mob is a fair fight
against the 3 starting characters. (I think, balance is hard to do
without testing and I am not sure how to get a fight started).